refactor: move from jest to vitest
