--- packages/octicons_react/jest.config.js
@@ -1,11 +1,0 @@
-'use strict'
-
-module.exports = {
-  moduleNameMapper: {
-    '^rollup$': require.resolve('rollup')
-  },
-  testEnvironment: 'jsdom',
-  transform: {
-    '^.+\\.(mjs|cjs|js|jsx)$': require.resolve('babel-jest')
-  }
-}
