+++ packages/octicons_react/package.json
@@ -23,10 +23,8 @@
     "build": "script/build.js && script/types.js && rollup -c",
     "clean": "rimraf .next dist src/__generated__",
     "develop": "next",
-    "lint": "eslint src pages script",
-    "test": "jest",
-    "ts-test": "tsc -P ts-tests",
-    "posttest": "npm run ts-test"
+    "lint": "eslint --ext .js,.jsx src pages script",
+    "ts-test": "tsc -P ts-tests"
   },
   "files": [
     "dist"
@@ -51,18 +49,14 @@
     "@rollup/plugin-commonjs": "22.0.2",
     "@rollup/plugin-node-resolve": "14.1.0",
     "@rollup/plugin-virtual": "2.1.0",
-    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^12.1.5",
     "@types/react": "^18.0.20",
-    "babel-jest": "29.0.3",
     "eslint": "^6.5.1",
     "eslint-plugin-github": "4.1.3",
-    "eslint-plugin-jest": "^21.17.0",
     "eslint-plugin-jsx-a11y": "6.4.1",
     "eslint-plugin-react": "7.23.2",
     "fs-extra": "^6.0.1",
-    "jest": "^29.0.3",
-    "jest-environment-jsdom": "29.0.3",
     "next": "^14.2.21",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
