--- packages/octicons_styled/babel.config.js
@@ -1,20 +1,0 @@
-'use strict'
-
-module.exports = {
-  presets: [
-    [
-      '@babel/preset-env',
-      {
-        targets: {
-          node: 'current'
-        }
-      }
-    ],
-    '@babel/preset-react'
-  ],
-  env: {
-    production: {
-      presets: ['next/babel']
-    }
-  }
-}
