--- packages/octicons_styled/jest.config.js
@@ -1,8 +1,0 @@
-'use strict'
-
-module.exports = {
-  testEnvironment: 'jsdom',
-  transform: {
-    '^.+\\.(mjs|cjs|js|jsx)$': require.resolve('babel-jest')
-  }
-}
