+++ packages/octicons_styled/package.json
@@ -16,9 +16,7 @@
   "scripts": {
     "build": "script/build.js && npm run rollup && script/copy.sh",
     "ts-test": "tsc -P ts-tests",
-    "test": "jest",
-    "posttest": "npm run ts-test",
-    "lint": "eslint src script",
+    "lint": "eslint --ext .js,.jsx src script",
     "rollup": "rollup -c --bundleConfigAsCjs"
   },
   "keywords": [
@@ -47,19 +45,14 @@
     "@github/prettier-config": "0.0.4",
     "@primer/octicons-react": "19.15.1",
     "@rollup/plugin-babel": "6.0.4",
-    "@testing-library/jest-dom": "^5.3.0",
+    "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^10.0.2",
     "@types/react": "^16.4.6",
     "eslint": "^6.5.1",
     "eslint-plugin-github": "4.1.3",
-    "eslint-plugin-jest": "^21.17.0",
     "eslint-plugin-jsx-a11y": "6.4.1",
     "eslint-plugin-react": "7.23.2",
     "fs-extra": "^6.0.1",
-    "babel-jest": "29.0.3",
-    "jest": "^29.0.3",
-    "jest-environment-jsdom": "29.0.3",
-    "jest-styled-components": "^7.0.2",
     "next": "^11.1.0",
     "react": "^16.4.0",
     "react-dom": "^16.4.1",
