+++ vitest.config.mjs
@@ -1,0 +1,7 @@
+import {defineConfig} from 'vitest/config'
+
+export default defineConfig({
+  test: {
+    projects: ['packages/*/vitest.config.mjs']
+  }
+})
